ENH: rename files and change structure
