+++ rocketpy/mathutils/function.py
