+++ rocketpy/motors/hybrid_motor.py
