+++ rocketpy/motors/motor.py
