+++ rocketpy/motors/solid_motor.py
